pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "miya-ai"
version = "4.0.0"
description = "弥娅AI虚拟形象系统"
readme = "README.md"
requires-python = ">=3.9"

# 核心依赖
dependencies = [
    # Web框架
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.22.0",
    "pydantic>=2.0.0,<3.0.0",
    "python-multipart>=0.0.6",
    "aiohttp>=3.8.0,<4.0.0",
    
    # 配置管理
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    
    # 系统工具
    "psutil>=5.9.0",
    "watchdog>=3.0.0",
    "requests>=2.31.0",
    
    # 异步文件操作
    "aiofiles>=23.0.0",
    
    # 安全/认证
    "PyJWT>=2.8.0",
    "cryptography>=41.0.0",
    
    # 类型提示
    "typing-extensions>=4.8.0",
]

# 入口点
[project.scripts]
miya = "run.main:main"
miya-terminal = "run.multi_terminal_main_v2:main"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--tb=short"